backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
torch>=2.2.0
torchvision>=0.17.0
albumentations>=1.4.0
safetensors>=0.4.0
opencv-python-headless>=4.9.0
huggingface-hub>=0.26.0
pillow>=10.0.0
numpy>=1.26.0
librosa>=0.10.0
soundfile>=0.12.0
scikit-learn>=1.4.0
tqdm>=4.66.0
